# refactor(rust): Dispatch new-streaming CSV source to updated multiscan