--- a/crates/polars-stream/src/nodes/io_sources/csv.rs
+++ b/crates/polars-stream/src/nodes/io_sources/csv.rs
@@ -140,7 +140,6 @@ impl FileReader for CsvFileReader {
             pre_slice,
             predicate: None,
             cast_columns_policy: _,
-            missing_columns_policy: _,
             num_pipelines,
             callbacks:
                 FileReaderCallbacks {