--- a/crates/polars-stream/src/nodes/io_sources/csv.rs
+++ b/crates/polars-stream/src/nodes/io_sources/csv.rs
@@ -670,7 +670,7 @@ impl ChunkReader {
                     ri.name.clone(),
                     ri.offset.clone(),
                     df.height(),
-                )?)
+                )?);
             }
         }