--- a/crates/polars-stream/src/nodes/io_sources/multi_scan.rs
+++ b/crates/polars-stream/src/nodes/io_sources/multi_scan.rs
@@ -75,6 +75,7 @@ pub struct MultiScanNode<T: MultiScanable> {
 
 impl<T: MultiScanable> MultiScanNode<T> {
     #[allow(clippy::too_many_arguments)]
+    #[expect(unused)]
     pub fn new(
         sources: ScanSources,