--- a/crates/polars-stream/src/physical_plan/lower_ir.rs
+++ b/crates/polars-stream/src/physical_plan/lower_ir.rs
@@ -478,6 +478,15 @@ pub fn lower_ir(
                     cloud_options,
                 )),
 
+                #[cfg(feature = "csv")]
+                FileScan::Csv {
+                    options,
+                    cloud_options,
+                } => Some((
+                    Arc::new(Arc::new(options.clone())) as Arc<dyn FileReaderBuilder>,
+                    cloud_options,
+                )),
+
                 #[cfg(feature = "json")]
                 FileScan::NDJson {
                     options,