--- a/crates/polars-stream/src/physical_plan/to_graph.rs
+++ b/crates/polars-stream/src/physical_plan/to_graph.rs
@@ -545,29 +545,7 @@ fn to_graph_rec<'a>(
                     #[cfg(feature = "ipc")]
                     polars_plan::dsl::FileScan::Ipc { .. } => unreachable!(),
                     #[cfg(feature = "csv")]
-                    polars_plan::dsl::FileScan::Csv {
-                        options,
-                        cloud_options,
-                    } => ctx.graph.add_node(
-                        nodes::io_sources::SourceComputeNode::new(
-                            nodes::io_sources::multi_scan::MultiScanNode::<
-                                nodes::io_sources::csv::CsvSourceNode,
-                            >::new(
-                                scan_sources.clone(),
-                                hive_parts.clone().map(Arc::new),
-                                *allow_missing_columns,
-                                include_file_paths.clone(),
-                                file_schema.clone(),
-                                projection.clone(),
-                                row_index.clone(),
-                                row_restriction.clone(),
-                                predicate,
-                                options.clone(),
-                                cloud_options.clone(),
-                            ),
-                        ),
-                        [],
-                    ),
+                    polars_plan::dsl::FileScan::Csv { .. } => unreachable!(),
                     #[cfg(feature = "json")]
                     polars_plan::dsl::FileScan::NDJson { .. } => unreachable!(),
                     _ => todo!(),
@@ -635,27 +613,7 @@ fn to_graph_rec<'a>(
                     #[cfg(feature = "ipc")]
                     FileScan::Ipc { .. } => unreachable!(),
                     #[cfg(feature = "csv")]
-                    FileScan::Csv { options, .. } => {
-                        assert!(predicate.is_none());
-
-                        if options.parse_options.comment_prefix.is_some() {
-                            // Should have been re-written to separate streaming nodes
-                            assert!(file_options.row_index.is_none());
-                            assert!(file_options.pre_slice.is_none());
-                        }
-
-                        ctx.graph.add_node(
-                            nodes::io_sources::SourceComputeNode::new(
-                                nodes::io_sources::csv::CsvSourceNode::new(
-                                    scan_source,
-                                    file_info,
-                                    file_options,
-                                    options,
-                                ),
-                            ),
-                            [],
-                        )
-                    },
+                    FileScan::Csv { .. } => unreachable!(),
                     #[cfg(feature = "json")]
                     FileScan::NDJson { .. } => unreachable!(),
                     _ => todo!(),