Also does some drive-by refactoring.

I have skipped quite a few tests that began failing due to a lack of a `missing_columns` parameter on the API - I plan to get to this later.

#### Benchmarks on 100x 1M CSV files (50 cols each):
Tested on AWS `c6gn.8xlarge`. Files are on local disk.

* `head(10_000_000)` (first 10 files)
```python
# Before (pypi 1.26.0)
30.17s user 6.52s system 872% cpu 4.205 total
# After
24.01s user 4.46s system 757% cpu 3.757 total
```

* `tail(10_000_000)` (last 10 files)
```python
# Before (pypi 1.26.0)
29.63s user 6.27s system 621% cpu 5.772 total
# After
25.36s user 5.00s system 778% cpu 3.901 total
```

@coastalwhite